Move Conversation and ConversationDetail components to lib/Scenes
