+++ src/lib/AppRegistry.tsx
@@ -7,7 +7,6 @@ import { SafeAreaInsets } from "lib/types/SafeAreaInsets"
 import { Theme } from "palette"
 import { ArtistQueryRenderer } from "./Containers/Artist"
 import { BidFlowQueryRenderer } from "./Containers/BidFlow"
-import { ConversationQueryRenderer } from "./Containers/Conversation"
 import { GeneQueryRenderer } from "./Containers/Gene"
 import { InboxQueryRenderer } from "./Containers/Inbox"
 import { InquiryQueryRenderer } from "./Containers/Inquiry"
@@ -26,6 +25,7 @@ import { CitySavedListQueryRenderer } from "./Scenes/City/CitySavedList"
 import { CitySectionListQueryRenderer } from "./Scenes/City/CitySectionList"
 import { CollectionQueryRenderer } from "./Scenes/Collection/Collection"
 import { CollectionFullFeaturedArtistListQueryRenderer } from "./Scenes/Collection/Components/FullFeaturedArtistList"
+import { ConversationQueryRenderer } from "./Scenes/Inbox/Screens/Conversation"
 
 // Consignments / My Collection
 import { Consignments } from "./Scenes/Consignments"
