+++ src/lib/Scenes/Inbox/Screens/Conversation.tsx
@@ -8,13 +8,13 @@ import React from "react"
 import { View } from "react-native"
 import { createFragmentContainer, graphql, QueryRenderer, RelayProp } from "react-relay"
 import styled from "styled-components/native"
-import ConnectivityBanner from "../Components/ConnectivityBanner"
-import Composer from "../Components/Inbox/Conversations/Composer"
-import Messages from "../Components/Inbox/Conversations/Messages"
-import { sendConversationMessage } from "../Components/Inbox/Conversations/SendConversationMessage"
-import { updateConversation } from "../Components/Inbox/Conversations/UpdateConversation"
-import { SmallHeadline } from "../Components/Inbox/Typography"
-import { Schema, Track, track as _track } from "../utils/track"
+import ConnectivityBanner from "../../../Components/ConnectivityBanner"
+import Composer from "../../../Components/Inbox/Conversations/Composer"
+import Messages from "../../../Components/Inbox/Conversations/Messages"
+import { sendConversationMessage } from "../../../Components/Inbox/Conversations/SendConversationMessage"
+import { updateConversation } from "../../../Components/Inbox/Conversations/UpdateConversation"
+import { SmallHeadline } from "../../../Components/Inbox/Typography"
+import { Schema, Track, track as _track } from "../../../utils/track"
 
 const Container = styled.View`
   flex: 1;
